+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.18] - 2022-06-02
 ### Fixed
+- Fix broken poll vote [#9](https://github.com/rokwire/polls-building-block/issues/9)
 - Additional fix for group admin should see all private polls [#9](https://github.com/rokwire/polls-building-block/issues/9)
 
 ## [1.0.17] - 2022-06-01